Move `TimeUtilities` to Explorer.Components namespace
Also remove unnecessary use of IAsynEnumerable.
